Fixed a high-FPS bug where the action step would be run independently of the process step

+++ src/Maverick/Entity/Object/Object.cpp
@@ -278,7 +278,7 @@ void Object::_process(uint64_t delta, unsigned int steps){
     }
     this->xA = 0;
 
-    //Updating Y values (CHANGE THESE TO ALTER VEL BY DELTA LATER)
+    //Updating Y values
     this->yV += this->yA;
     if(this->yV != 0){
         this->yV -= this->yV * friction;
@@ -300,7 +300,7 @@ void Object::_process(uint64_t delta, unsigned int steps){
 
     //Updating Z values
     if(this->z != this->ground && this->gravity){
-        this->zA -= engine->getGravity() * steps;
+        this->zA -= engine->getGravity() * (float)steps;
     }
 
     this->zV += this->zA;
